Remove .idea and CMakeLists.txt files from the repository
IDE files as a general software engineering practice are not
version controlled. Furthermore, the IDEs update them whenever
the code is changed. This has the effect of drowning actual code
changes in commit diffs. A few lines of code changes entail many
more lines of IDE files' changes.

CMakeLists.txt is superfluous. R has a command to build packages which
we have been using successfully so far. Having another build system
will increase maintenance burden. Every time we create a new file, we
have to add it to CMakeLists.txt. We also have to ensure that code
compiled with CMake is passed the same set of flags as passed when
building with R. These flags may not be the same on all systems. This
invites opportunities for bugs.