Add code that creates OK files on successful tracing
If the tracing is successful, a ".OK" file is created
to identify databases which are safe for analysis.
Earlier, this file was being generated externally by
the instrumented vignette. However, that would always
create a .OK file, even if the tracer was unsuccessful
in tracing the vignette. This commit adds code to the
tracer to create this file in the `end` function. This
function only gets called if the tracing is successful.