Add serialization of return value type of function calls
This commit adds a new table to the database which relates
the type of the return value of each function call with the
call id. This information is required by some analyses, for
e.g. to find out functions that return environments. These
functions can potentially be involved in implementing lazy
data structures.