Fix implementation of SEXP type handling
This commit replaces the poorly done sexp_type. A lot of
unnecessary code was written to handle conversion of sexp_type
with some un-handled cases and conversion bugs.
This commit removes all this redundancy and solves the bugs as
a consequence by replacing the sexp_type enum with sexptype_t
typedef.